Bulk org/org-roam PKMS overhaul
This is a large set of changes to PKMS functionality in general.
It was necessary to develop the whole thing as a unit like this, but I
do not have the energy to refactor all this crap into separate atomic
commit. Whoever finds this will just have to deal with it. I am sorry.