Make org-agenda time-grid look more consistent
When using different types of logged entries, the time-grid shows them
in different formats. For example, with CLOCK: entries it shows a (-)
before the timestamp, and with :LOGBOOK: State changes it shows the
state. None of them are aligned, and the time-grid stops being a
"grid". It doesn't make sense and my ADHD can't handle it and I hate
it.